# prowl/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prowl LLM Orchestration API",
    "description": "REST API for LLM-driven orchestration of the Prowl security crawler. Enables programmatic crawl control, discovery queries, mid-crawl adjustments, intervention handling, and attack surface reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/endpoints": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Query discovered endpoints",
        "description": "Returns endpoints with rich filtering. Supports filtering by tag, page_type, HTTP method, auth requirement, parameter presence, source module, and URL pattern.",
        "operationId": "list_endpoints_api_v1_endpoints_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "description": "Page size (max 500)",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size (max 500)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag (e.g. api, admin, debug)",
              "title": "Tag"
            },
            "description": "Filter by tag (e.g. api, admin, debug)"
          },
          {
            "name": "page_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by page type (login, error, api_json, admin, etc.)",
              "title": "Page Type"
            },
            "description": "Filter by page type (login, error, api_json, admin, etc.)"
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by HTTP method (GET, POST, PUT, etc.)",
              "title": "Method"
            },
            "description": "Filter by HTTP method (GET, POST, PUT, etc.)"
          },
          {
            "name": "requires_auth",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by auth requirement",
              "title": "Requires Auth"
            },
            "description": "Filter by auth requirement"
          },
          {
            "name": "has_params",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parameter presence",
              "title": "Has Params"
            },
            "description": "Filter by parameter presence"
          },
          {
            "name": "source_module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source module (s1_spider, s4_js, etc.)",
              "title": "Source Module"
            },
            "description": "Filter by source module (s1_spider, s4_js, etc.)"
          },
          {
            "name": "url_pattern",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on URL",
              "title": "Url Pattern"
            },
            "description": "Substring match on URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/graph": {
      "get": {
        "tags": [
          "v1",
          "discovery"
        ],
        "summary": "Get site structure graph",
        "description": "Returns a node/edge graph of discovered paths for visualization.",
        "operationId": "get_graph_api_v1_graph_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Maximum number of endpoints to include",
              "default": 500,
              "title": "Limit"
            },
            "description": "Maximum number of endpoints to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Graph Api V1 Graph Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/interventions/{intervention_id}/resolve": {
      "post": {
        "tags": [
          "v1",
          "control"
        ],
        "summary": "Resolve an intervention",
        "description": "Resolve an intervention by providing auth data (cookies, headers, token, credentials). The engine is automatically resumed after resolution.",
        "operationId": "resolve_intervention_api_v1_interventions__intervention_id__resolve_post",
        "parameters": [
          {
            "name": "intervention_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Intervention Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveInterventionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperationResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Operation result: ok, paused, resumed, stopped, resolved, error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OperationResponse"
      },
      "PaginatedResponse": {
        "properties": {
          "items": {
            "items": {},
            "type": "array",
            "title": "Items",
            "description": "Page of results"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total matching items"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Current offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether more items exist beyond this page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse",
        "description": "Cursor-based pagination wrapper."
      },
...
      "ResolveInterventionRequest": {
        "properties": {
          "cookies": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Cookies",
            "description": "Session cookies"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers",
            "description": "Auth headers"
          },
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Bearer token",
            "default": ""
          },
          "credentials": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Credentials",
            "description": "Login credentials (username, password)"
          },
          "extra_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra Data",
            "description": "Additional resolution data"
          }
        },
        "type": "object",
        "title": "ResolveInterventionRequest",
        "description": "Provide auth data to resolve an intervention."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}